+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* Set datetime formater textfield
+* Set formater for `type1`, `type2`, `type3`, `type4`.
+* Add value notifier for datetime